Fix schedtool -E on i386
The sched_setattr() macro currently discards the "flags" parameters.
If function parameters are passed in the same registers used for Linux
syscalls (as in the x86_64 ABI), this does not cause any issue, because
the "flags" value is already in the correct register. But if function
parameters are passed on the stack (as in the i386 ABI), the sched_setattr()
syscall ends up being called with a random third parameter and fails.